requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
azure-common==1.1.26
    # via azure-storage
azure-nspkg==3.0.2
    # via
    #   azure-common
    #   azure-storage
azure-storage==v0.20.3
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
django-storages==1.4.1
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
django==1.11.29
    # via -r requirements/base.in
futures==3.3.0 ; python_version == "2.7"
    # via
    #   -c requirements/constraints.txt
    #   azure-storage
idna==2.10
    # via requests
python-dateutil==2.8.1
    # via azure-storage
pytz==2020.5
    # via
    #   -r requirements/base.in
    #   django
requests==2.25.1
    # via azure-storage
six==1.15.0
    # via python-dateutil
urllib3==1.26.2
    # via requests
